Extract robotics synchronization helpers
